+++ AI/roop/requirements.txt
@@ -18,3 +18,4 @@ opennsfw2==0.10.2
 protobuf==4.23.4
 tqdm==4.65.0
 python-dotenv~=1.0.1
+celery~=5.3.6
